Organizing some code.
